Update `snapshot_property` table tests: adjust column definitions, add new constraints for primary key, foreign key, and unique index; update `instruction_index` data type in related tests.

+++ test/LogixDb.Data.SqlServer.Tests/Migrations/M20260306Tests.cs
@@ -15,7 +15,7 @@ public class M20260306Tests : SqlServerTestFixture
             await AssertColumnDefinition("instruction", "instruction_id", "int");
             await AssertColumnDefinition("instruction", "snapshot_id", "int");
             await AssertColumnDefinition("instruction", "rung_hash", "nvarchar");
-            await AssertColumnDefinition("instruction", "instruction_index", "int");
+            await AssertColumnDefinition("instruction", "instruction_index", "smallint");
             await AssertColumnDefinition("instruction", "instruction_key", "nvarchar");
             await AssertColumnDefinition("instruction", "instruction_text", "nvarchar");
             await AssertColumnDefinition("instruction", "is_conditional", "bit");
